MLK-20518: hdp: Fix memory out of bounds access
Fix memory out of bounds access.
Change arry type for functopn avi info frame,
Align the arry type and its length.

(cherry picked from commit 2fc41a88c9da514ae3f377e7cb73f4df886f038e)

+++ drivers/mxc/hdp/API_Infoframe.h
@@ -53,11 +53,11 @@
  * \{
  */
 CDN_API_STATUS CDN_API_InfoframeSet(state_struct *state, u8 entry_id,
-				    u8 packet_len, u32 *packet,
+				    u8 packet_len, u8 *packet,
 				    u8 packet_type);
 CDN_API_STATUS CDN_API_InfoframeSetNoActiveIdle(state_struct *state,
 						u8 entry_id, u8 packet_len,
-						u32 *packet, u8 packet_type);
+						u8 *packet, u8 packet_type);
 CDN_API_STATUS CDN_API_InfoframeRemove(state_struct *state, u8 entry_id);
 CDN_API_STATUS CDN_API_InfoframeRemovePacket(state_struct *state, u8 entry_id, u8 packet_type);
